Rename mock_adapter to common_mocks and move it to mocks folder
