+++ test/unit/adapter/test_adapt_registry.py
@@ -7,7 +7,7 @@ import pytest
 from golem.core.adapter import register_native, AdaptRegistry
 from golem.core.optimisers.fitness import SingleObjFitness, Fitness
 from golem.core.optimisers.graph import OptGraph, OptNode
-from test.unit.adapter.mock_adapter import MockAdapter, MockDomainStructure, MockNode
+from test.unit.mocks.common_mocks import MockAdapter, MockDomainStructure, MockNode
 from test.unit.utils import graphs_same
