+++ test/unit/serialization/test_decoder.py
@@ -4,7 +4,7 @@ import pytest
 from golem.serializers import CLASS_PATH_KEY, Serializer
 from .dataclasses.serialization_dataclasses import DecoderTestCase
 
-from .mocks.serialization_mocks import MockGraph, MockNode, MockOperation
+from test.unit.mocks.serialization_mocks import MockGraph, MockNode, MockOperation
 from .shared_data import (
     MOCK_NODE_1,
     MOCK_NODE_2,
